+++ NEWS.md
@@ -1,5 +1,12 @@
 # News for Package maotai
 
+### changes in version 0.1.5
+  * Following functions are added,
+    - `bmds`     : Bayesian Multidimensional Scaling.
+    - `cmds`     : Classical Multidimensional Scaling.
+    - `kmeanspp` : k-means++ algorithm for clustering.
+    - `tsne`     : t-Stochastic Neighbor Embedding.
+
 ### changes in version 0.1.4
   * Update `README` for listing the available functions in the package.
   * Corrected an Armadillo type-casting error.